feat: animação do jogador, zoom na câmera e chão placeholder

+++ src/modules/animation.h
@@ -13,8 +13,8 @@ typedef struct anim {
   usint num_frames;   // Quantos frames tem a animação
   usint loop_frame;   // Para qual frame a animação loopa ao chegar no final
   bool loop;          // Se a animação loopa ou não
-  double timer;       // Tempo decorrido desde a última troca de frames
-  double frame_dur;   // Duração de um frame da animação
+  float timer;        // Tempo decorrido desde a última troca de frames
+  float frame_dur;    // Duração de um frame da animação
   Vector2 frame_size; // Altura e largura de cada frame
 } Animation;
